dlcv-fall-2025-hw2-FrankLin-123712/requirements.txt
numpy==1.24.4
torch==1.11.0
torchvision==0.12.0
timm==0.8.19.dev0
pandas==2.0.3
Pillow==10.4.0
scipy==1.9.1
scikit-image==0.20.0
albumentations==0.4.3
opencv-python==4.1.2.30
pudb==2019.2
invisible-watermark
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
torch-fidelity==0.3.0
transformers==4.46.3
torchmetrics==0.6.0
kornia==0.6
matplotlib==3.7.0
detectors==0.1.8
gdown==5.2.0
tqdm
detectors
-e git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
-e git+https://github.com/openai/CLIP.git@main#egg=clip
